requirements.txt
biopython==1.74
primer3-py==0.6.0
networkx==2.3
